backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.10
langchain-community>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
numpy>=2.0.0
httpx>=0.25.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
